+++ griffon/commands/queries.py
@@ -274,7 +274,7 @@ def retrieve_component_summary(ctx, component_name, strict_name_search, operatio
     "search_all",
     is_flag=True,
     default=False,
-    help="Search all Components and dependencies.",
+    help="Flat search for all Components.",
 )
 @click.option(
     "--search-all-roots",
@@ -282,13 +282,6 @@ def retrieve_component_summary(ctx, component_name, strict_name_search, operatio
     is_flag=True,
     default=False,
     help="Search all ROOT Components and dependencies.",
-)
-@click.option(
-    "--search-redhat",
-    "search_redhat",
-    is_flag=True,
-    default=False,
-    help="Search all Red Hat root Components and dependencies.",
 )
 @click.option(
     "--search-community",
@@ -302,7 +295,7 @@ def retrieve_component_summary(ctx, component_name, strict_name_search, operatio
     "search_all_upstreams",
     is_flag=True,
     default=False,
-    help="Search for Components by upstream.",
+    help="Flat search for all Components by upstream.",
 )
 @click.option(
     "--no-community",
@@ -396,7 +389,6 @@ def get_product_contain_component(
     filter_rh_naming,
     search_all,
     search_all_roots,
-    search_redhat,
     search_community,
     search_all_upstreams,
     no_community,
@@ -423,7 +415,6 @@ def get_product_contain_component(
         and not search_related_url
         and not search_community
         and not search_all_upstreams
-        and not search_redhat
         and not search_provides
         and not search_upstreams
     ):